rename indicator.js -> create-element.js
